Remove `__doc` flag in favor of auto-detection
docs.rs has an option to use Cargo metadata to pass additional flags.
This can be used for docs.rs, but building documentation for
time-rs.github.io needs the flags passed via an environment variable.
This is done in CI.

As a result of this change, CI can now use `--all-features` on stable
without issue. This also has an impact on IDEs (for the better).